music-preprocessing/requirements-preprocessing.txt
# Audio & ML stack
audioread==3.0.1
bottleneck
brotli
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama
contourpy
cycler==0.11.0
decorator==5.2.1
fonttools
idna==3.10
importlib-resources
joblib==1.4.2
kiwisolver
lazy-loader==0.4
librosa==0.11.0
llvmlite
matplotlib==3.5.3
mkl-service==2.4.0
mkl-fft
mkl-random
msgpack==1.1.0
numba
numexpr
pandas
pillow
platformdirs==3.11.0
pooch==1.8.2
pycparser==2.22
pyparsing
PyQt6==6.7.1
PyQt6-sip
python-dateutil
pytz
requests==2.32.3
resampy
scikit-learn==1.2.2
scipy
setuptools==72.1.0
sip
six==1.16.0
soundfile==0.13.1
soxr
threadpoolctl
tornado
tqdm
typing-extensions==4.13.2
tzdata
unicodedata2
urllib3
wheel==0.45.1
zipp

